Move install_pypi_records & as_explicit
