+++ conda_lockfiles/create.py
@@ -8,9 +8,20 @@ from conda.misc import explicit
 
 from .exceptions import LockfileFormatNotSupported
 from .loaders import LOADERS
-from .utils import as_explicit, install_pypi_records
+
+import os
+import sys
+from shutil import which
+from subprocess import run, CompletedProcess
+from tempfile import NamedTemporaryFile
+from typing import TYPE_CHECKING
+
+from conda.common.compat import on_win
+from conda.models.records import PackageRecord
+from conda.models.prefix_graph import PrefixGraph
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
     from conda.common.path import PathType
     from pathlib import Path
 
@@ -36,3 +47,45 @@ def create_environment_from_lockfile(
         if verbose:
             print("Installing PyPI packages:")
         install_pypi_records(pypi, prefix)
+
+
+def install_pypi_records(pypi_records: Iterable[str], prefix: str) -> CompletedProcess:
+    if not pypi_records:
+        return
+    with NamedTemporaryFile("w", delete=False) as f:
+        f.write("\n".join(pypi_records))
+
+    if executable := which("uv"):
+        command = [executable]
+    else:
+        command = [sys.executable, "-m"]
+    if on_win:
+        python_exe = os.path.join(prefix, "python.exe")
+    else:
+        python_exe = os.path.join(prefix, "bin", "python")
+    command += [
+        "pip",
+        "install",
+        "--prefix",
+        prefix,
+        "-r",
+        f.name,
+        "--no-deps",
+        "--python",
+        python_exe,
+    ]
+    try:
+        return run(command, check=True)
+    finally:
+        os.unlink(f.name)
+
+
+def as_explicit(
+    records: Iterable[PackageRecord],
+    **comments,
+) -> Iterable[str]:
+    for key, value in comments.items():
+        yield f"# {key}: {value}"
+    yield "@EXPLICIT"
+    for record in dict.fromkeys(PrefixGraph(records).graph):
+        yield f"{record.url}#{record.sha256}"
